requirement.txt
crewai
openai
langchain
chromadb
pdfplumber
python-docx
tiktoken
python-dotenv
docling
tenacity